Move {Setup, Rollover}Params to model module
They are obviously part of the model, since they are the input to
their respective protocols.
